Introduce Import::Commitments
This is a class to import Commitments into the application.

The motivation is to include a commitment value that can then be
exported in the IATI XML to improve BEIS Publish What You Fund score.

Right now the intention is that it will be run by a developer so it does
very little authorisation, it can also import to activities at any level
and for any organisation as this is what we believe the is required to
meet the BEIS user need.

Like most importers, it is difficult to not end up finding and updating
for each import row. I did try Rails 6 `upsert` and `upsert_all` but in
this context, the benefit didn't seem worthwhile – this is pretty much
a one time import for the moment so performance is not as important
here.

The import will create commitments for activites which do not have one
and overwrite the `value` for any that do.

The import also runs all changes in a transaction and if ANY row cannot
be imported, all changes are rolled back – this is the standard approach
across the application.

I decided to load csv fixture files as I feel like this is good approach
for an importer that imports csv files from disk.

I also introduce a `import` name space in services, I feel like as we
have multiple importers it would be good to try and collect them
together and perhaps even begin to functionality.