Reset benefitting_countries when both legacy region and country were set
It seems we were over-eager in backfilling the new `Activity#benefitting_countries`
field, in that we added all of the countries in a legacy #recipient_region even when
recipient_country was present. (We weren't expecting that both of these could be set.)

 e.g.
```
activity = Activity.find_by(roda_identifier: 'GCRF-RFNetG-R2-GCRFNGR2\10190')
activity.attributes.select { |k,v| k.match(/(geo)|(count)|(bene)|(region)/) }

   => {"recipient_region"=>"1030",
    "recipient_country"=>"NG",
    "geography"=>"recipient_region",
    "intended_beneficiaries"=>[],
    "country_delivery_partners"=>[],
    "benefitting_countries"=>
     ["NG", "BJ", "BF", "CV", "CI", "GM", "GH", "GN", "GW", "LR", "ML", "MR",
      "NE", "SH", "SN", "SL", "TG"]}
```

To remedy, we find activities where:

- legacy region is present, and
- legacy country is present, and
- the newly populated #benefitting_countries is greater than
  the total legacy countries (#intended_beneficiaries.count + 1 #recipient_country )

and we reset the new #benefitting_countries field to the content of legacy
recipient_country and the #intended_beneficiaries.

Note: we use `COALESCE()` to cast null values in array fields to zero.