Don't return headers if there are no forecasts
We discovered a bug when downloading a
spending breakdown for an organisation / fund
that has no forecasts set. This results in the
following error:

```
ArgumentError: wrong number of arguments (given 1, expected 2..3)
```

This is because when we try to build the
headers for forecasts, we call `Range.new`
with the first argument being the financial
quarter after the last one we have actuals for
(`all_actual_and_refund_financial_quarter_range.last.succ`)
and the second argument being the `max`
value of all the Financial Quarters we have
forecasts for
(`all_financial_quarters_with_forecasts.max`).

If we have no forecasts, this argument is `nil`,
and that's what gives us the error.

To get round this, we check if
`all_financial_quarters_with_forecasts` is
blank first (i.e. an empty array). If it is, we
return an empty array for the range. I've also
removed the splat (`*`) from the second
argument, as it didn't offer any value.