Include commitment in the IATI XML
The commitment is a transaction on the activity with `transaction-type`
of 2.

We do not include as much information on a commitment as other types of
transactions.

We do not include the currency as leaving it off implies the activity
default currency which we do include.

The transaction date and the value date are both derived from the
commitment financial quarter, we use the first day of that financial
quarter as the more specific date required.

See IATI Transaction docs:

https://iatistandard.org/en/iati-standard/203/activity-standard/iati-activities/iati-activity/transaction/

A note about the Bullet safe list:

We use a generic service to find activities of a certain level and we
try in to `include` all the related entities, including a Commitment.

Bullet is watching for cases where these includes are not used, but
as not all activities will have a Commitment, there are valid cases when
the include is not required.

I don't feel like handling these situations in the finders is the right
solution as they would soon become difficult to interact with, having to
specify which associations to include each use.

As the Commitment is genuinely optional I think it is safe add it to the
safe list.