Raise error from CreateRefund when invalid
We've experienced a few occurences of a bug where the refund
form was submitted in the shape of `params[:refund]` rather
than the expected shape of `params[:refund_form]`.

This bug is a hangover from when the  `RefundForm` was
introduced to handle the complexity of validating an entity
composed of both a Refund and a Comment.

The bug occurs in the scenarios where a Refund can not
actually be persisted by the CreateRefund service and is
wrongly passed back to the view for correction.

A overview of the classes involved:

- `RefundForm`: a "form object" which facilitates validation
of both `Refund` and `Comment` objects

- `RefundController`: passes the RefundForm to and from the
view layer, including for correction of validation errors
and for editing

- `CreateRefund`: service which automatically assigns a report
and attempts to persist a `Refund` record

If it's impossible for the `CreateRefund` service to actually
persist the intended refund, this is almost certainly because
an editable report can't be found and assigned automatically
by the service.

This isn't a problem which the user can fix by changing the
form inputs so it's wrong to handle this scenario as a
validation problem to be passed back to the controller and
view.

We have good test coverage on the `RefundPolicy` and
`ActivityPolicy#create_refund` themselves, and at present we
don't know which validation errors the `CreateRefund`
services has passed back to the controller and view, which were then
mishandled causing the form to be resubmitted in the shape
of `params[:refund]` rather than `params[:refund_form]`.

The changes here mean that future occurrences will be
surfaced clearly as errors, and we will them be able to fix
the defective behaviour leading to those persistence
failures.