Add Active Record attribute loacles
We have a wider plan to make these strings the canonical source for the
strings used to identify attributes.

This is a first step on that path and does not try to be perfect, it
gives enough strings for all the attributes I came up with by running:

`Activity.first.attributes.keys`

The included strings are the one agreed be BEIS and may not align to
various places in the application - wider work is planned to fix that.