Introduce ActivityAttributesColumns
This is the first in a series of new models that allow the application
to export data in a more modular way.

This first class can be used to export the non-finance activity data and
sets the pattern for others to follow.

Passing an array of activities and an array of attributes will give you
access to:

- an array of the Headers from the locales
- a hash of the activity values keyed by activity id

The data structure can then be used to assemble a csv export.

The activity id is used to assemble data from other classes that will
follow this one, for example combining all non-financial data with net
actual spend totals and forecasts totals to produce the report csv.

Passing the attributes in a defined order gives you them back in that
order, which allows us to set the order in a logical way.

The Activity models contains attributes that we do not want to surface
to users so there is an IGNORED_ATTRIBUTES list and attribute on the list
are simply not included in either the header or row output.

Finally passing an unknown attribute will result in a
UnknownAttributeError as this would be the same error raised if
you called an unknown attribute directly.