Group comments by their associated activity
Now all comments have a report, we need to
fetch the assocaited activity to group them by.
In the case of comments made directly against
an activity, this is the Activity itself, but for
Refunds and Adjustments, it will be the
associated activity.

I did consider adding another direct
association, in a similar way to how we made
the report association mandatory. However, as
we're not directly accessing the activity from
the database, this seems like an uneccesary
overhead.

To get the associated activity, I've added a
new `associated_activity` method. I've been
very prescriptive about fetching the activity,
because even though we could do a simple
`if` statement checking if the `commentable`
is an Activity, this could cause problems
further down the line if we choose to add a
new commentable type, so it's better to be
prescriptive.