Introduce ActivityActualsColumns
Returns the actuals financial columns.

Extracted out of the existing Export::SpendingBreakdown class.

Actuals are made up of:

- Actual spend + any actual adjustments
- Refunds + any refund adjustments
- Net spend (the total of the first two added together)

The class produces the standard version by default (or by setting
`include_breakdown` to false), where you will get a single net spend total
column for each financial quarter

Setting `include_breakdown` to true will give you all three columns for each
financial quarter, see above.

The default behaviour is to return the actuals for all time i.e all
actuals, refunds and adjustments in the database when the code is run,
as the data appears in the spending breakdown report.

By passing in a report you can get a snapshot in time of the values up
to and including the report, this will not include any adjustments made
after the report.

The report behaviour is provided by the existing AllActivityTotals
class, so is no exhaustively tested at this level, we rely on the only
totals available to parse into columns to be the correct ones, here we
are mostly looking at the rendered column headers and value are as
expected.