Introduce ActivityForecastColumns
This is extracted from `Export::SpendingBreakdown` so that more exports
can make use of the same behaviour.

At this point it behaves like the spending breakdown, it returns all
available forecast values for the activities. The values are the latest
version the db contains see Forecast docs for more information:

https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/blob/develop/doc/forecasts-and-versioning.md