Update the report comment view
This updates the comment reports view to
show more detail about a comment, including
the type. Because adjustment and refund
comments can't be edited, I've added two
simple policies that just return `false` for the
`update?` action. This seems simpler than
adding a bunch of conditionals, and also
allows us to build on these policies if
necessary further down the line.