ActivityDeliveryPartnerOrganisationColumn
This class gets the delivery partner organisation for activities.

If the activity is a level B (Programme) activity it uses the
Implementing organisation, of which there can be multiple, delimiting
them with a pipe (|).

For all other levels the `organisation` association is used.

To optimise the performance of an instance of this class  it must be
passed an ActiveRecord:Relation for the internal eager loading to work
correctly, as it chains onto the passed activities.