Introduce ActivityChangeStateColumn
This requires a `Report` in order to work out if change has occurred, the
report gives the context in order to do so.

I am sticking with passing in the activities even though we could use
`all_report_activities`, if the activity is not in the report we return
nil - I think this is preferable over throwing an error that may not be
immediately clear.