Ensure adjustment relates to an earlier (than this report's) time period
We have validation in 2 places:

1. primarily mainly in the `AdjustmentForm`
2. also in the `CreateAdjustment` service where the
   association with the report is made

We now handle (2) properly by initialising a fresh
`AdjustmentForm` if the form view needs to be re-rendered
for the user to choose a valid report, or else consider a
different way of making the correction.

Perhaps we ought to refactor this for the `AdjustmentForm` to
handle the assignment of the report. At present the
CreateAdjustment service handles 3 error scenarios:

1. report isn't editable: raises error as it should not
have been possible to fill in the form. It could be dealing
correctly with a race condition, e.g. where the report has
just been submitted by someone else

2. the report is not associated with the given activity:
raises an error, as this really should not happen.

3. the report is for the same financial period as the
adjustment: n this case we want to re-render the form and
explain to the user that they should either:

   a) *edit* the relevant actual or refund if the financial
      period is the current one. Or they should

   b) choose a previous financial period if the correction
      relates to a mistake or omission in a previously approved
      report and an Adjustment is indeed the appropriate
      correction mechanism.