Rails validation to ensure no unapproved reports
There is already a database constraint to ensure there is only one editable report per organisation and fund.

As part of moving towards manually created reports, we are surfacing this constraint at the Rails model validation, to make it easier to surface error messages to the users.

The difference between the database constraint and the Rails validation, which is the presence of an `”inactive”` state, will go away in future commits where we will be removing the inactive state from the state machine.