Allow setting the report financial quarter and year
In order to allow BEIS users to create reports manually, we need to allow them to set the financial quarter and year.

To avoid breaking all the existing specs and to keep the commits easier to read and review, we are still setting the FQ/FY automatically if no values have been provided for them.