Fix forecast policy
Delivery partner users cannot access another organisation’s forecasts in the first place, but the policy was theoretically allowing them to update or destroy forecasts belonging to any organisation, as long as there was an editable report for that org.

This wasn’t previously surfaced by specs, because the reports were created as inactive by default, so the specs passed thanks to the editable report condition rathen than organisation belonging.

The changes to the spec are only to improve performance by reusing the expensive setup for each context.