Allow fetching comments for a specific activity
Though we introduced the concept of an
associated_activity, we still need to add a
somewhat complex query with some joins to
fetch all comments relating to an activity. I feel
this is a fair tradeoff as is reduces complexity
and duplication elsewhere.

I've split it into two scopes - the first:
`with_commentables` does the necessary
joins to make the `for_activity` query possible.

We can then use the `for_activity` scope in the
Activity's `has_many` association, making
sure we use `unscope` to make sure the
default scope is not included.