Set report on all comments
Currently, the `report` association is optional.
To display all comments made during a
reporting period without having to make
massively complex queries, it's best if we set
the report on a comment every time one is
created.

With this in mind, I've remvoed the `optional`
flag from the model, and updated the
`CreateAdjustment` and `CreateRefund`
services to add the report at the time the
associated comment is created.