Do not set FQ and FY automatically on report
Because we’re removing the mechanism to set the financial quarter and year on reports automatically, we need to validate their presence for any newly created reports.

We still need to allow historical reports to be valid without FQ/FY, and we need to allow them to be created programmatically in the specs and via console.