Extend Export::ActivityActualsColumns
We extend this to take an optional `report`.

The default behaviour is to return the actuals for all time i.e all
acutals, refunds and adjustments in the database when the code is run.

By passing in a report you can get a snapshot in time of the values up
to and including the report, this will not include any adjustments made
after the report.