Get a report 'snapshot' of Forecasts
The reason we store versions of Forecast is to allow a user to view a
snapshot of the forecasts in a given report.

Here we pass in a Report and show the forecast values at that point in
time i.e the forecast value as they were when that report was made.