Allow a starting financial quarter
In the spending breakdown export users want the latest forecasted values
but only for the financial quarters that do not have actual spend.

To achieve this we allow a FinancialQuarter object to be passed in. The
language can get a bit confusing here, a FinancialQuarter object is
actually made up of a financial_quarter and financial_year which is what
we want.

The forecast values are ALWAYS the latest version, if you want past
values, pass in a report.

We don't want to pass both a report and a starting financial quarter as
passing a report also sets the starting financial quarter, so we raise
an argument error if this happens.