Introduce ActivityImplementingOrganisationColumn
Allows Implementing organisations to be exported.

Returns the name only as the only use case we have for this
right now is the report csv which only requires the name.

Multiple implementing organisations are delimited with a pipe (|).

When the activity is at Level B (a programme) there is no guarantee
there will be any implementing organisations, but the
extending organisation will have the correct organisation details, so we
use that.

We pass in an ActiveRecord::Relation of activities so we can leverage
Rails lazy loading of the organisations, failing to pass in the correct
type would cause an error, failing to pass in the correct type would
cause an error, so we check this.