Track changes when a forecast is changed
This records a historical event when a forecast
is changed by a delivery partner, outside of its
original reporting cycle. If a forecast is created
and then edited in the same reporting cycle, or
created/edited by a service owner (who do not
do these things during reporting cycles), we
do nothing.

I've also attempted to clean up and rationalise
the forecast history specs to make them easier
to follow. Whether or not this has worked is
another matter entirely!