Improve SpendingBreakdown spec
We can use the same inline csv fixtures to make this spec a little more
approachable.

As the spending breakdown now uses the ActivityActualsColumn, which is
tested, we can ease the specs here, we only want confidence that the
export returns values, we get confidence from ActivityActualsColumn spec
that the values are correct.

I've added some additional coverage to ensure the latest forecast values
are included, although this is a bit ugly right now, this should all
disappear once we extract the forecast calculation from here into it's
own object.