Make sure financial year gets selected properly
Before, the financial year on the budget form
was getting reset on edit each time as
`financial_year` returns a `FinancialYear`
object, which gets transformed to a string by
default, which doesn't translate to any item
in the list of options. Using `to_i` returns an
integer, which means the correct item gets
selected.

I've also added `include_blank: true` to make
sure that setting the financial year is a
deliberate act, rather than accidentally
choosing the first item in the list.