Record a historical event when editing a budget
This hooks into `UpdateBudget` to record any
changes that are made to a budget by a
delivery partner. I've used some of the magic
contained in `ActiveModel::Dirty` to track the
changes becuase it's a useful thing to hook
into, and also we use enums for some fields, so
it's easier and cleaner to track the actual old
and new value, rather than the integer-based
value.