Record change to Budget for all users
Previously only delivery partner users triggered the addition of a
historical event when a budget attribute was changed.

We expect BEIS users can make changes to budgets outside of a Report and
to have those changes recorded in the history, when a change is made by
a BEIS user we expect the HistoricalEvent.report to be nil.

We had a duplicate `budget_type` attribute and I cannot see the value to
end users of recording a change of `providing_organisation_id` as this
won't mean anything to them so I have removed both from the list of
attribute changes we record a historical event for.

We add some coverage around the history changes being recorded correctly
here as well.