Do not cache Budget#financial_year
Overriding and memoising the result of this attribute accessor could
cause unexpected results.

For example, the `UpdateBudget` service loads a Budget object and sets
it's new attributes, a developer would expect to set a new
`financial_year` attribute here and see the change, but would not.

The only way to see the change would be to initalise a new instance of
the Budget.

There may have been a thought that Budget financial years should not be
change, but I don't think that is the case and I don't think this would
be the approach to take if it were.