Change benefitting_countries_with_percentages
The old calculation using `divmod` fails for numbers > 100.

As we have ~180 countries and selecting all of them is something that
happens we need the percentage calculation to work.

What matters to us right now is that the IATI XML is valid and NOT that
the percentage is accurate (it is not and for BEIS is may be impossible
to make it accurate).

I tested what percentages can be valid in the IATI XML for both 3
countries and 180 and this simpler calculation satisfies the validator.