Add a delete activity rake task
Every report results in a bunch of activities that a developer has to
delete as the application does not support deleting activities.

This rake task attempts to aid developers in performing this support
task by running the steps for a given activity database ID.

Some associations are destroyed along with the Activity object, others
would be orphaned. I wanted to update the documentation we have on this
and through the process of finding out, I figured we may as well wrap
the knowledge in a rake task.

Developers are encouraged to think about what they are doing here as
ultimately this is destructive, but they save a great deal of work on
the console by running the task and interpreting the result before performing
the destruvtive action. Please enter the commit message for your changes.