New variance column
This replicates the variance calculation as used in the report export.

You take the current report financial quarter forecast value and take
away the net actual spend for the same financial quarter.

At this point this is the simplest version of this for use along with
other `Export` classes.

`Export::ActivityActualsColumns` provides `rows_for_last_financial_quarter`
and `Export::ActivityForecastcolumns` provides
`rows_for_first_financial_quarter` to pass into this along with the
activities and the expected financial quarter.

From this the variance is calculated.

I decided to pass in the activities as I feel like I want it to remain
explicit that we are still using a set of activities to generate the
rows to export.

I also decided to pass in the financial quarter as this class is
accepting row data only which has not concept of the financial quarter
it represents, further refactoring could make this better, but that is
not for now - this simple version gets us where we need to be.