Expose first forecast and last actual column data
Variance is the difference between the forecasted spend and the actual
spend. Here we allow the `Export::ActivityActualsColumns` to expose it's
last financial quarter of data and `Export::ActivityForecastColumns` to
expose it's first.

Right now, we only show variance in the 'report csv' which is always
tied to a report, this means we only ever want these columns of data to
calculate variance as the last column of actuals will be the same
financial quarter as the report and the first forecast financial quarter
will be the same.

The actual calculation and resulting column comes next.