Tie existing actuals and forecast to explicit financial period and report
We'll go on to assert that only forecasts for periods on
which we have not yet reported should be totalled up in:

"Total forecasted spend"

i.e. we have 2 forecasts:

1. Q3 2021: £40
2. Q4 2021: £30

Given we have reported for Q3 2021, our "Total forecasted"
should be £30, the total of later forecasts.

NB: Bullet prompted us to add an "include" from `Forecast` to
`:parent_activity` in the `Tab` helper class.