Cache the #rows in ActivityActualsColumns
Like most `Export` classes, ActivityActualsColumns represents the data
it is initialised with as columns in an export, here we cache the result
of the `rows` method so the expensive database operation is only called
once for each activity.

Any further calls will use the cache, which is fine for this class as you
cannot change the data it represents without initialising  a new
instance.

This will become very useful in coming work where we will need to
extract the last column of data for further processing.