Introduce ActivityCommentsColumn
This class is concerned with exporting a column of activity comments for
a report.

As there are an unknown number of comments that must be added
into a single table cell, so we output the comments separated with the
following:

`----`

which is surrounded by new lines.

We would usually delimit with a pipe but that did not feel like enough
(currently comments are separated by a new line).