Model the BEIS defined order of fields
BEIS feel there is an order to the fields in the report csv, here we
create a basic class that models this.

The order here from this work:

https://trello.com/c/bf9t1om8

Supplied by BEIS. Not all the fields in the supplied order will appear
here as some are not attributes of the `Activity` model.