Add actuals columns to report export
At this point we change the way we assemble the columns. This is because
in the rare case that there are no Actuals there is no column to render,
we can only check this in the `Export::Report` as this is where the
assembly takes place.

We are only asserting that the correct columns are in place here as we
have confidence that the column classes themselves are returning the
correct values.