Cannot download a report with no financial quarter
The interesting change here is around a report that has no financial
quarter. Historically we have two of these reports for each delivery
partner, they were used early on to import historic data.

To support rendering these reports would mean adding a great deal of
complexity to the Export classes as we rely on the report financial
quarter to determine which columns and which versions of data to show.

The application has moved on a long way from when supporting these
reports with nil financial quarters was necessary and I believe the
exports meet any user need that this functionality met in the past. The
spending breakdown report includes ALL acutal spend include any that
would have been imported in these two reports.

I think there is an argument to add a 'fake' financial quarter to these
reports, they were performed in a financial quarter after all, but I
will leave that as separate work - the real value here is the new report
export that includes many improvements.

I've also updated a couple of feature specs for the new report export.

I've dropped the file name assertions as the methods that produce the
string are tested in the class itself - we simply assert we get a
success code and that it is csv file type - this feels like the right
level for a feature spec to me.