Prevent negative actuals unless importing history
Actuals should never be negative unless the users is BEIS and is
importing historic actuals (which are allowed to be negative as the
effort to fix them is seen to be too great).

Here we apply the validation on `Actual` unless the `importing_history`
validation context is passed. This validation is not used yet, that is
to follow.