New Actual history importer
This class imports supplied actuals in the `history` context, i.e. the
value can be negative, but not zero.

It accepts a `Report` as nothing should be added without and a
`CSV::Table` of the data - this is the expectation we have for all our
imports and the output of `CsvFileUpload.rows`.

We check for valid headers first and bail on the import if they cannot
be found, not having all the correct headers would result in a failed
import for ALL rows.

For each row we check the report is editable and that the RODA ID
results in an activity, this two items are required to make a valid
`Actual`.

If everything is good, we create a new `Actual` and validate it with the
`history` context and save it.

If any other attributes are not considered valid for the activity, we
pass those errors back up as `RowErrors` so they can be displayed to the
user.

All rows are checked so that a user can understand and fix all the
errors in the file at one time.