Extract RowError
Our users import lots of data from csv, it would be good for us to
handle all errors from this process the same way, here I extract the
`RowError` class into it's own file and extend it.

`RowError` is already used in `Import::Commitments` and
`Import::ActualHistory` but does not store all of the information we
want to help users who are importing data:

- the column header that the error relates to
- the row number in the csv file that contains the error
- the value that triggered the error
- a message to describe the error

As a subclass of Ruby's `StandardError` I couldn't see the value of
testing it, I did write a spec, but it seemed too simple to be of value.

RowError is not fussy and will happily except `nil`. This is expected as
some errors do not relate to a particular column or row, a header
problem for instance.

Use Import::RowError in the actual history importer lays the ground work
to display import errors to users.