Define column names in locales
We are trying to consolidate the places where we define column names in
the import files, export files and the application interface, this will
make for a more consistent user experience.

Here we can use the already defined 'RODA identifier' locale on
`Activity`.

Financial quarter, Financial year and Value are attributes on multiple
models so we cannot rely on just one to provide the strings. I've
decided to use the `default` locale scope for these as a sensible place
to store these kind of locales that relate to multiple models.