Show unknown country codes
We have had a lot of churn in the country codes in the data set.

Whilst we already show the user when there is an unknown country code we
do not tell them what the code is. This information is very useful in
'debugging' the data and deciding on a course of action.