Data migration to remove country codes
We discovered we have ended up with these two incompatible country
codes. Once is not a code at all and the other is a real country code
but should not be used in BEIS ODA management at this time.

We are only removing codes from the current benefitting countries field.