Activity summary view users BEIS locale strings
We are slowing refactor where and what we use to describe the various
attributes in use throughout the service.

With this change the Activity summary view and report csv file share the
same locales.

Here the Activity summary view is updated to use the
`activerecord.attributes.activity` namespace whenever possible. Not
everything is an activerecord attributes so I have had to use my best
judgement when to include a key and when to not.