Add basic Sidekiq stats to health check endpoint
We want to know if the Sidekiq worker has failed for any reason and raise
an alert.

We decided it was acceptable to add some very basic stats to the public
`health_check` endpoint for this purpose.

We leverage the Sidekiq API to return the two statistics we want:

- the size of `enqueued`
- the `retry_size`

See:

https://github.com/mperham/sidekiq/wiki/API

and

https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/blob/develop/doc/background-jobs.md