Migrate unique implementing orgs to organisations table
We know from IATI that Organisations can play multiple roles:

- **Funding**: The government or organisation which
  provides funds to the activity.

- **Accountable**: An organisation responsible for
  oversight of the activity and its outcomes

- **Extending**: An organisation that manages the budget
  and direction of an activity on behalf of the funding
  organisation

- **Implementing**: The organisation that physically carries out
  the activity or intervention.

At present RODA uses 2 db tables:

1.   `organisations` for "extending organisations", aka
"delivery partners"

2. `implementing_organisations` for "implementing
organisations"

There are two duplication issues:

i. Due to the uncontrolled way they are managed, there are currently an
ever increasing number of duplicate implementing organisations, created
with slightly different names.

ii. Many of the extending organisations (delivery partners) also have
entries in the implementing organisation table

In this commit we create listings in the _organisations_
table for all of the unique implementing organisations not
represented.

As `Activity#implementing_organisations` does not use a
'join' table to support the fact that implementing
organisations have many activites, there are many
duplicated and near-duplicate entries. To handle this we
add a **temporary** `Organisation#alternate_names`
column to handle the variants in naming seen in the
`implementing_organisations` table.

We'll use these alternate names when we come to create
organisation memberships for each of the entries in the
`implementing_organisations` table.